Added file
Cloned from Marker.Tracking.Test this file detects specific markers NOT object display to determine what is in camera view. The component is attached to each marker and uses an event listener to update feedback.

Next steps are to determine best way to check for all Three markers being visible to trigger and event to hide all objects and display a new object. Possible a customEvent will do the trick.